+++ src/Command/Crawler/AbstractCrawler.php
@@ -10,12 +10,13 @@ namespace HeimrichHannot\NewsBundle\Command\Crawler;
 
 use Contao\NewsModel;
 use GuzzleHttp\Client;
+use Symfony\Component\Console\Style\SymfonyStyle;
 
 abstract class AbstractCrawler implements CrawlerInterface
 {
     const ERROR_NO_ERROR = 0;
     const ERROR_BREAKING = 1;
-    const ERROR_NOTICE = 2;
+    const ERROR_NOTICE   = 2;
 
     /**
      * @var Client
@@ -39,21 +40,29 @@ abstract class AbstractCrawler implements CrawlerInterface
      * @var array
      */
     protected $error = [
-        'code' => 1,
+        'code'    => 1,
         'message' => 'No error specified'
     ];
+
+    /**
+     * @var null|SymfonyStyle
+     */
+    protected $io = null;
 
     /**
      * AbstractCrawler constructor.
      * @param Client $client
      * @param NewsModel $item
      * @param string $baseUrl
+     * @param null|SymfonyStyle $io
      */
-    public function __construct($client, $item = null, $baseUrl = '')
+    public function __construct($client, $item = null, $baseUrl = '', $io = null)
     {
         $this->client  = $client;
         $this->item    = $item;
         $this->baseUrl = $baseUrl;
+        $this->count   = 0;
+        $this->io      = $io;
     }
 
     /**
@@ -63,19 +72,36 @@ abstract class AbstractCrawler implements CrawlerInterface
     public function getUrls()
     {
         \System::getContainer()->get('contao.framework')->initialize();
-        $urls   = [];
+        $urls = [];
         if (isset($GLOBALS['TL_HOOKS']['addNewsArticleUrlsToSocialStats'])
             && is_array($GLOBALS['TL_HOOKS']['addNewsArticleUrlsToSocialStats'])) {
             foreach ($GLOBALS['TL_HOOKS']['addNewsArticleUrlsToSocialStats'] as $callback) {
                 $addUrls = \System::importStatic($callback[0])->{$callback[1]}($this->item, $this->baseUrl);
                 $urls    = array_merge(
-                    $addUrls,
+                    is_array($addUrls) ? $addUrls : [],
                     $urls
                 );
             }
         }
+
+        // handle umlauts, spaces etc
+        foreach ($urls as $i => $url) {
+            $urls[$i] = rawurldecode($url);
+        }
+
+        if ($this->io) {
+            $this->io->text($urls);
+        }
+
         return $urls;
     }
+
+    /**
+     * Return a count for the current item at crawler provider or an error.
+     *
+     * @return int|array
+     */
+    abstract function getCount();
 
     /**
      * Update the current item
@@ -116,13 +142,13 @@ abstract class AbstractCrawler implements CrawlerInterface
         $this->baseUrl = $baseUrl;
     }
 
-    public function setErrorCode ($code)
+    public function setErrorCode($code)
     {
         $this->error['code'] = $code;
         return $this->error;
     }
 
-    public function setErrorMessage ($message)
+    public function setErrorMessage($message)
     {
         $this->error['message'] = $message;
         return $this->error;
@@ -136,7 +162,21 @@ abstract class AbstractCrawler implements CrawlerInterface
         return $this->error;
     }
 
+    /**
+     * @return null|SymfonyStyle
+     */
+    public function getIo(): SymfonyStyle
+    {
+        return $this->io;
+    }
 
+    /**
+     * @param null|SymfonyStyle $io
+     */
+    public function setIo($io)
+    {
+        $this->io = $io;
+    }
 
 
 }